Move TM2 dataRequest out of main manager
