+++ traffic_monitor/experimental/traffic_monitor/enum/enum.go
@@ -6,6 +6,8 @@ import (
 	"strings"
 )
 
+type TrafficMonitorName string
+
 type CacheName string
 
 type CacheGroupName string
