+++ traffic_monitor/experimental/traffic_monitor/manager/healthresult.go
@@ -2,15 +2,17 @@ package manager
 
 import (
 	"fmt"
+	"log"
+	"sort"
+	"sync"
+	"time"
+
 	"github.com/Comcast/traffic_control/traffic_monitor/experimental/traffic_monitor/cache"
 	ds "github.com/Comcast/traffic_control/traffic_monitor/experimental/traffic_monitor/deliveryservice"
 	"github.com/Comcast/traffic_control/traffic_monitor/experimental/traffic_monitor/enum"
 	"github.com/Comcast/traffic_control/traffic_monitor/experimental/traffic_monitor/health"
 	"github.com/Comcast/traffic_control/traffic_monitor/experimental/traffic_monitor/peer"
 	todata "github.com/Comcast/traffic_control/traffic_monitor/experimental/traffic_monitor/trafficopsdata"
-	"log"
-	"sync"
-	"time"
 )
 
 type DurationMapThreadsafe struct {
@@ -54,26 +56,24 @@ func (o DurationMapThreadsafe) Set(cacheName string, d time.Duration) {
 
 // StartHealthResultManager starts the goroutine which listens for health results.
 // Returns the last health durations, events, and the local cache statuses.
-func StartHealthResultManager(cacheHealthChan <-chan cache.Result, toData todata.TODataThreadsafe, localStates CRStatesThreadsafe, statHistory StatHistoryThreadsafe, monitorConfig TrafficMonitorConfigMapThreadsafe, peerStates CRStatesPeersThreadsafe, combinedStates CRStatesThreadsafe) (DurationMapThreadsafe, EventsThreadsafe, CacheAvailableStatusThreadsafe, DSStatsThreadsafe, StatsLastKbpsThreadsafe) {
+func StartHealthResultManager(cacheHealthChan <-chan cache.Result, toData todata.TODataThreadsafe, localStates CRStatesThreadsafe, statHistory StatHistoryThreadsafe, monitorConfig TrafficMonitorConfigMapThreadsafe, peerStates CRStatesPeersThreadsafe, combinedStates CRStatesThreadsafe, fetchCount UintThreadsafe, errorCount UintThreadsafe) (DurationMapThreadsafe, EventsThreadsafe, CacheAvailableStatusThreadsafe, DSStatsThreadsafe, StatsLastKbpsThreadsafe) {
 	lastHealthDurations := NewDurationMapThreadsafe()
 	events := NewEventsThreadsafe()
 	localCacheStatus := NewCacheAvailableStatusThreadsafe()
 	dsStats := NewDSStatsThreadsafe()
 	lastKbpsStats := NewStatsLastKbpsThreadsafe()
-	go healthResultManagerListen(cacheHealthChan, toData, localStates, lastHealthDurations, statHistory, monitorConfig, peerStates, combinedStates, events, localCacheStatus, dsStats, lastKbpsStats)
+	go healthResultManagerListen(cacheHealthChan, toData, localStates, lastHealthDurations, statHistory, monitorConfig, peerStates, combinedStates, fetchCount, errorCount, events, localCacheStatus, dsStats, lastKbpsStats)
 	return lastHealthDurations, events, localCacheStatus, dsStats, lastKbpsStats
 }
 
-func healthResultManagerListen(cacheHealthChan <-chan cache.Result, toData todata.TODataThreadsafe, localStates CRStatesThreadsafe, lastHealthDurations DurationMapThreadsafe, statHistory StatHistoryThreadsafe, monitorConfig TrafficMonitorConfigMapThreadsafe, peerStates CRStatesPeersThreadsafe, combinedStates CRStatesThreadsafe, events EventsThreadsafe, localCacheStatus CacheAvailableStatusThreadsafe, dsStats DSStatsThreadsafe, lastKbpsStats StatsLastKbpsThreadsafe) {
-	errorCount := uint64(0) // TODO make atomic shared with manager.go
+func healthResultManagerListen(cacheHealthChan <-chan cache.Result, toData todata.TODataThreadsafe, localStates CRStatesThreadsafe, lastHealthDurations DurationMapThreadsafe, statHistory StatHistoryThreadsafe, monitorConfig TrafficMonitorConfigMapThreadsafe, peerStates CRStatesPeersThreadsafe, combinedStates CRStatesThreadsafe, fetchCount UintThreadsafe, errorCount UintThreadsafe, events EventsThreadsafe, localCacheStatus CacheAvailableStatusThreadsafe, dsStats DSStatsThreadsafe, lastKbpsStats StatsLastKbpsThreadsafe) {
 	lastHealthEndTimes := map[string]time.Time{}
 	healthHistory := map[string][]cache.Result{}
-	fetchCount := uint64(0) // TODO change to atomic/mutex in manager.go
 	eventIndex := uint64(0) // TODO move to EventsThreadsafe.Add() ?
 	for {
 		select {
 		case healthResult := <-cacheHealthChan:
-			fetchCount++
+			fetchCount.Inc()
 			toDataCopy := toData.Get() // create a copy, so the same data used for all processing of this cache health result
 			var prevResult cache.Result
 			if len(healthHistory[healthResult.Id]) != 0 {
@@ -101,7 +101,7 @@ func healthResultManagerListen(cacheHealthChan <-chan cache.Result, toData todat
 			var err error
 			newDsStats, newLastKbpsStats, err := ds.CreateStats(statHistory.Get(), toDataCopy, combinedStates.Get(), lastKbpsStats.Get(), now)
 			if err != nil {
-				errorCount++
+				errorCount.Inc()
 				log.Printf("ERROR getting deliveryservice: %v\n", err)
 			} else {
 				dsStats.Set(newDsStats)
@@ -116,4 +116,41 @@ func healthResultManagerListen(cacheHealthChan <-chan cache.Result, toData todat
 			lastHealthEndTimes[healthResult.Id] = now
 		}
 	}
+}
+
+// calculateDeliveryServiceState calculates the state of delivery services from the new cache state data `cacheState` and the CRConfig data `deliveryServiceServers` and puts the calculated state in the outparam `deliveryServiceStates`
+func calculateDeliveryServiceState(deliveryServiceServers map[string][]string, states CRStatesThreadsafe) {
+	deliveryServices := states.GetDeliveryServices()
+	for deliveryServiceName, deliveryServiceState := range deliveryServices {
+		if _, ok := deliveryServiceServers[deliveryServiceName]; !ok {
+			// log.Printf("ERROR CRConfig does not have delivery service %s, but traffic monitor poller does; skipping\n", deliveryServiceName)
+			continue
+		}
+		deliveryServiceState.IsAvailable = false
+		deliveryServiceState.DisabledLocations = nil
+		for _, server := range deliveryServiceServers[deliveryServiceName] {
+			if states.GetCache(server).IsAvailable {
+				deliveryServiceState.IsAvailable = true
+			} else {
+				deliveryServiceState.DisabledLocations = append(deliveryServiceState.DisabledLocations, server)
+			}
+		}
+		deliveryServices[deliveryServiceName] = deliveryServiceState
+	}
+	states.SetDeliveryServices(deliveryServices)
+}
+
+// intersection returns strings in both a and b.
+// Note this modifies a and b. Specifically, it sorts them. If that isn't acceptable, pass copies of your real data.
+func intersection(a []string, b []string) []string {
+	sort.Strings(a)
+	sort.Strings(b)
+	var c []string
+	for _, s := range a {
+		i := sort.SearchStrings(b, s)
+		if i < len(b) && b[i] == s {
+			c = append(c, s)
+		}
+	}
+	return c
 }
